+++ frontend/src/app/apps/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import Apps from '@/app/apps/_components/Apps/Apps';
+import Apps from '@/app/apps/_components/UserApps/UserApps';
 
 export default function App() {
 	return <Apps />;
