+++ frontend/src/app/loading.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import styles from '@/styles/loading.module.css';
+
+export default function Loading() {
+	return <h1 className={styles.loading}>Loading...</h1>;
+}
