+++ frontend/src/styles/loading.module.css
@@ -1,0 +1,8 @@
+.loading {
+    display: inline-flex;
+    align-items: center;
+    margin: 0 auto;
+    text-align: center;
+    width: 100%;
+    height: 100%;
+}
